BACKPORT: userns: When the per user per user namespace limit is reached return ENOSPC
The current error codes returned when a the per user per user
namespace limit are hit (EINVAL, EUSERS, and ENFILE) are wrong.  I
asked for advice on linux-api and it we made clear that those were
the wrong error code, but a correct effor code was not suggested.

The best general error code I have found for hitting a resource limit
is ENOSPC.  It is not perfect but as it is unambiguous it will serve
until someone comes up with a better error code.

Change-Id: I570756ee15eb28416ce68d218f8815ccd61f47b2
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
(cherry picked from commit df75e7748bae1c7098bfa358485389b897f71305)