BACKPORT: net: Add a xfrm validate function to validate_xmit_skb
When we do IPsec offloading, we need a fallback for
packets that were targeted to be IPsec offloaded but
rerouted to a device that does not support IPsec offload.
For that we add a function that checks the offloading
features of the sending device and and flags the
requirement of a fallback before it calls the IPsec
output function. The IPsec output function adds the IPsec
trailer and does encryption if needed.

Change-Id: I493a16019d53d8215cc420398367abd3b50078c0
Signed-off-by: Steffen Klassert <steffen.klassert@secunet.com>
(cherry picked from commit f6e27114a60a0afdec40db1bf7f6da37b565745a)